Refactor: 필터링 전체로 묶고 컴포넌트로 따로 뺌(#15)
